\name{twowaytests-package}
\alias{twowaytests-package}
\alias{twowaytests-package}
\docType{package}
\title{
Two-Way Tests in Independent Groups Designs
}
\description{
Performs two-way tests in independent groups designs. These are two-way ANOVA, two-way ANOVA under heteroscedasticity: parametric bootstrap based generalized test and generalized pivotal quantity based generalized test (Ananda et al., 2022) <doi:10.1080/03610926.2022.2059682>, two-way ANOVA for medians, trimmed means, M-estimators (Wilcox, 2011; ISBN:978-0-12-386983-8). The package performs descriptive statistics and graphical approaches. Moreover, it assesses variance homogeneity and normality of data in each group via tests and plots. All 'twowaytests' functions are designed for two-way layout.
}
\details{
\tabular{ll}{
Package: \tab twowaytests\cr
Type: \tab Package\cr
License: \tab GPL (>=2)\cr
}
}