\name{twowaytests-package}
\alias{twowaytests-package}
\alias{twowaytests-package}
\docType{package}
\title{
Two-Way Tests in Independent Groups Designs
}
\description{
Performs two-way tests in independent groups designs; Parametric Bootstrap based Generalized Test and Generalized Pivotal Quantity based Generalized Test (Weerahandi and Krishnamoorthy, 2019) <doi:10.1080/03610926.2017.1419264>. The package performs descriptive statistics and graphical approaches. Moreover, it assesses variance homogeneity and normality of data in each group via tests and plots. All 'twowaytests' functions are designed for two-way layout.
}
\details{
\tabular{ll}{
Package: \tab twowaytests\cr
Type: \tab Package\cr
License: \tab GPL (>=2)\cr
}
}